db/requirements.txt
psycopg2-binary==2.9.9
Faker==25.2.0
python-dateutil==2.9.0.post0